chore(pipeline): extract helper types to separate files
